feat(ma): add catalog src in workspace root, move calendar to src and change name calendar to ma-calendar
