+++ micro-apps/package.json
@@ -57,8 +57,8 @@
     "lint:code:fix": "eslint --config ./@linter/eslint.config.cjs . --fix --ext \".js, .ts\"",
     "lint:code:run": "eslint --config ./@linter/eslint.config.cjs . --ext \".js, .ts\"",
     "lint:staged:fix": "lint-staged --config ./@linter/lint-staged.config.cjs",
-    "lint:styles:fix": "stylelint --config ./@linter/stylelint.config.cjs \"calendar/**/*.scss\" --fix",
-    "lint:styles:run": "stylelint --config ./@linter/stylelint.config.cjs \"calendar/**/*.scss\"",
+    "lint:styles:fix": "stylelint --config ./@linter/stylelint.config.cjs \"src/**/*.scss\" --fix",
+    "lint:styles:run": "stylelint --config ./@linter/stylelint.config.cjs \"src/**/*.scss\"",
     "lint:types:run": "tsc",
     "test:ui:run": "playwright test -c ./@test/playwright/playwright-ct.config.mjs",
     "test:ui:watch": "playwright test -c ./@test/playwright/playwright-ct.config.mjs --ui",
