+++ micro-apps/src/ma-calendar/src/domain/contracts/repository.contract.ts
