+++ micro-apps/src/ma-calendar/src/domain/contracts/service.contract.ts
