+++ micro-apps/src/ma-calendar/src/domain/models/days.model.ts
