+++ micro-apps/src/ma-calendar/src/ui/modules/days/days.module.scss
