+++ micro-apps/src/ma-calendar/src/ui/modules/days/month-days/month-days.module.scss
