+++ micro-apps/src/ma-calendar/src/ui/modules/days/week-days/week-days.module.scss
