+++ micro-apps/src/ma-calendar/src/ui/modules/month-switcher/month-switcher.tsx
