+++ micro-apps/src/ma-calendar/src/ui/modules/month-switcher/use-month-switcher.hook.ts
