+++ micro-apps/src/ma-calendar/src/ui/modules/wrapper/wrapper.module.scss
