+++ micro-apps/src/ma-calendar/src/ui/modules/year-switcher/use-year-switcher.hook.ts
