+++ micro-apps/src/ma-calendar/src/ui/modules/year-switcher/year-switcher.tsx
