+++ micro-apps/src/ma-calendar/src/ui/views/calendar.view.tsx
